+++ cub3d/includes/settings.h
@@ -6,8 +6,8 @@
 
 # define WIN_WIDTH 1000
 # define WIN_HEIGHT 600
-# define MAX_MAP_WIDTH WIN_WIDTH 
-# define MAX_MAP_HEIGHT WIN_HEIGHT 
+# define MAX_MAP_WIDTH WIN_WIDTH / 4 
+# define MAX_MAP_HEIGHT WIN_HEIGHT / 4 
 # define PLAYER_SPEED 1
 # define PLAYER_COLOR COLOR_RED
 //# define PI 3.14159265359
