+++ cub3d/includes/types.h
@@ -6,7 +6,7 @@
 /*   By: znajdaou <znajdaou@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/07 14:13:23 by znajdaou          #+#    #+#             */
-/*   Updated: 2025/08/05 15:09:41 by znajdaou         ###   ########.fr       */
+/*   Updated: 2025/08/07 17:18:05 by znajdaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,6 +41,22 @@ typedef struct s_map{
   int **arr;
 } t_map;
 
+//##################################
+//# raycasting
+//##################################
+typedef struct s_ray {
+  t_cor hitv; // hit point in the vertical lines
+  t_cor hith; // hit point in the horizontal lines
+  int dh; // if the ray facing the down it's 1, if up its -1 
+  int dv; // if the ray facing the right it's 1, if up its -1
+  double distv;
+  double disth;
+} t_ray;
+
+
+//##################################
+//# Obejcts
+//##################################
 // rectongle
 typedef struct s_rect
 {
@@ -65,8 +81,6 @@ typedef struct s_line
   t_cor e; // end
 	int color;
 }	t_line;
-
-
 
 
 // player
